MAJOR CHANGE: Complete AppMainTree!

+++ src/routes/app/(app)/+server.ts
@@ -30,9 +30,9 @@ const handlePOST = async (request: PostRequest): Promise<_Response> => {
     case RequestType.Thread:
       return handlePostThreadReq(request.data);
     case RequestType.ThreadFolder:
-      return handlePostThreadFolder(request.data);
+      return handlePostThreadFolderReq(request.data);
     case RequestType.PromptFolder:
-      return handlePostPromptFolder(request.data);
+      return handlePostPromptFolderReq(request.data);
     default:
       throw new Error("ERROR ON SERVER SIDE");
   }
@@ -47,15 +47,22 @@ const handleDELETE = async (request: DeleteRequest) => {
     case RequestType.Thread:
       handleDeleteThreadReq(data); // TODO: return item was removed successfully?
       break;
+    case RequestType.ThreadFolder:
+      handleDeleteThreadFolderReq(data); // TODO: return item was removed successfully?
+      break;
+    case RequestType.PromptFolder:
+      handleDeletePromptFolderReq(data); // TODO: return item was removed successfully?
+      break;
     default:
       throw new Error("Error while handling POST request");
   }
 };
 
+const PostEntriesSchema = z.object({ name: z.string().min(1) });
+
 // IMPURE CODE:
-const NewPromptEntriesSchema = z.object({ name: z.string().min(1) });
 const handlePostPromptReq = async (data: NewPrompt): Promise<_Response> => {
-  const entriesCheck = NewPromptEntriesSchema.safeParse(data);
+  const entriesCheck = PostEntriesSchema.safeParse(data);
 
   if (!entriesCheck.success) {
     return {
@@ -65,15 +72,15 @@ const handlePostPromptReq = async (data: NewPrompt): Promise<_Response> => {
   }
 
   const prompt: Prompt = await promptDAO.insert(data);
-  promptFoldersDAO.addItem(data.parentId, prompt._id);
+  if (data.parentId !== "") {
+    promptFoldersDAO.addItem(data.parentId, prompt._id);
+  }
 
   return { success: true, data: prompt };
 };
 
-const NewThreadEntriesSchema = z.object({ name: z.string().min(1) });
-
 const handlePostThreadReq = async (data: NewThread): Promise<_Response> => {
-  const entriesCheck = NewThreadEntriesSchema.safeParse(data);
+  const entriesCheck = PostEntriesSchema.safeParse(data);
 
   if (!entriesCheck.success) {
     return {
@@ -83,43 +90,64 @@ const handlePostThreadReq = async (data: NewThread): Promise<_Response> => {
   }
 
   const thread: Thread = await threadDAO.insert(data);
-  threadFoldersDAO.addItem(data.parentId, thread._id);
+
+  if (data.parentId !== "") {
+    threadFoldersDAO.addItem(data.parentId, thread._id);
+  }
 
   return { success: true, data: thread };
 };
 
-const handlePostThreadFolder = async (data: NewFolder): Promise<_Response> => {
-  console.log("TODO: handlePostThreadFolder");
-  const folder: Folder = {
-    _id: "TODO",
-    parentId: "TODO",
-    name: data.name,
-  };
+const handlePostThreadFolderReq = async (data: NewFolder): Promise<_Response> => {
+  const entriesCheck = PostEntriesSchema.safeParse(data);
+
+  if (!entriesCheck.success) {
+    return {
+      success: false,
+      issues: entriesCheck.error.errors.map((err) => err.message), // NOTE: Bude musete být key-value, aby bylo možno vyhodnotit, jaká hodnota byla v dialogu chybně zadána
+    };
+  }
+
   return {
     success: true,
-    data: folder,
+    data: await threadFoldersDAO.insert(data),
   };
 };
 
-const handlePostPromptFolder = async (data: NewFolder): Promise<_Response> => {
-  console.log("TODO: handlePostPromptFolder");
-  const folder: Folder = {
-    _id: "TODO",
-    parentId: "TODO",
-    name: data.name,
-  };
+const handlePostPromptFolderReq = async (data: NewFolder): Promise<_Response> => {
+  const entriesCheck = PostEntriesSchema.safeParse(data);
+
+  if (!entriesCheck.success) {
+    return {
+      success: false,
+      issues: entriesCheck.error.errors.map((err) => err.message), // NOTE: Bude musete být key-value, aby bylo možno vyhodnotit, jaká hodnota byla v dialogu chybně zadána
+    };
+  }
+
   return {
     success: true,
-    data: folder,
+    data: await promptFoldersDAO.insert(data),
   };
 };
 
 const handleDeletePromptReq = (data: DeleteRequestData) => {
   promptDAO.deleteById(data._id);
-  promptFoldersDAO.removeItem(data.parentId, data._id);
+  if (data.parentId !== "") {
+    promptFoldersDAO.removeItem(data.parentId, data._id);
+  }
 };
 
 const handleDeleteThreadReq = (data: DeleteRequestData) => {
   threadDAO.deleteById(data._id);
-  threadFoldersDAO.removeItem(data.parentId, data._id);
+  if (data.parentId !== "") {
+    threadFoldersDAO.removeItem(data.parentId, data._id);
+  }
+};
+
+const handleDeleteThreadFolderReq = (data: DeleteRequestData) => {
+  threadFoldersDAO.deleteById(data._id);
+};
+
+const handleDeletePromptFolderReq = (data: DeleteRequestData) => {
+  promptFoldersDAO.deleteById(data._id);
 };
